refactor(ffi-bridge): split error/vault.rs into directory module (closes #44)
The bridge `error/vault.rs` had grown to 702 LOC — 202 over the 500-line
file-split threshold. Issue #44 recorded the obligation and named the
preferred shape: `error/vault/{mod.rs, tests.rs}` over `#[path]`, so the
codebase keeps a single mod-file convention.

Move the `#[cfg(test)] mod tests` body into a sibling `tests.rs`,
declared from `mod.rs` as `#[cfg(test)] mod tests;`. The enum definition
and `From<core::VaultError>` impl stay together in `mod.rs` (they share
exhaustive variant coverage; splitting them would weaken the typed-error
discipline). External callers reach `FfiVaultError` via `pub mod vault;`
re-exported by the parent `error/mod.rs` — directory-module conversion
is path-invisible, so no call sites churn.

Resulting line counts:
- `error/vault/mod.rs`: 347 LOC (enum + From impl)
- `error/vault/tests.rs`: 363 LOC (24 behavior-pinning tests)

Pure refactor. No public-API change. Test counts identical
(642 cargo + 9 ignored), clippy / fmt / conformance / spec-freshness
all PASS.
